Extract methods from ellipse and resolve markup issue
