+++ src/annotations/markups/ellipse/ellipse.js
@@ -1,134 +1,12 @@
-import Style from "ol/style/Style";
-import Stroke from "ol/style/Stroke";
-import dcmjs from "dcmjs";
-import { fromCircle } from "ol/geom/Polygon";
-import CircleGeometry from "ol/geom/Circle";
-import { Feature } from "ol";
-import CircleStyle from "ol/style/Circle";
-import Point from "ol/geom/Point";
-import Fill from "ol/style/Fill";
-import RegularShape from "ol/style/RegularShape";
-
 import Enums from "../../../enums";
 import { distance } from "../bidirectional/mathUtils";
-import { getFeatureScoord3dArea } from "../../../scoord3dUtils";
-import { getUnitSuffix } from "../../../utils";
 import annotationInterface from "../../annotationInterface";
 import { getEllipseHandlesId, getEllipseId } from "./id";
-
-const styles = {
-  image: {
-    radius: 7,
-  },
-  stroke: {
-    color: "black",
-    width: 3,
-  },
-};
-
-const getEllipseStyle = () =>
-  new Style({
-    stroke: new Stroke({
-      color: styles.stroke.color,
-      width: styles.stroke.width,
-    }),
-  });
-
-const getHandleStyle = (coords) =>
-  new Style({
-    geometry: new Point(coords),
-    image: new CircleStyle({
-      radius: styles.image.radius,
-      stroke: new Stroke({
-        color: styles.stroke.color,
-        width: styles.stroke.width,
-      }),
-    }),
-  });
-
-function ellipseHandlesStyleFunction(feature) {
-  const styles = [];
-  const featureGeometry = feature.getGeometry();
-
-  const type = featureGeometry.getType();
-  if (type !== "LineString") {
-    return styles;
-  }
-
-  const [start, end] = featureGeometry.getCoordinates();
-
-  featureGeometry.forEachSegment((seg) => {
-    if (
-      seg[0] !== start[0] ||
-      seg[0] !== start[1] ||
-      seg[1] !== start[0] ||
-      seg[1] !== start[1]
-    ) {
-      const hiddenPointStyle = new Style({
-        geometry: new Point(seg),
-        image: new RegularShape({
-          fill: new Fill({
-            color: "rgba(255, 255, 255, 0.5)",
-          }),
-          stroke: new Stroke({
-            color: "red",
-            width: 1,
-          }),
-          points: 4,
-          radius: 8,
-          angle: Math.PI / 4,
-        }),
-      });
-      styles.push(hiddenPointStyle);
-    }
-  });
-
-  styles.push(getHandleStyle(start));
-  styles.push(getHandleStyle(end));
-
-  return styles;
-}
-
-const updateMarkup = (
-  feature,
-  { markupManager, drawingSource, map, pyramid }
-) => {
-  const ellipseFeatureId = getEllipseId(feature);
-  const ellipseFeature = drawingSource.getFeatureById(ellipseFeatureId);
-  if (ellipseFeature) {
-    const view = map.getView();
-    const unitSuffix = getUnitSuffix(view);
-    const ellipseGeometry = ellipseFeature.getGeometry();
-    const area = getFeatureScoord3dArea(ellipseFeature, pyramid);
-
-    let sum = 0;
-    let count = 0;
-    let sumSquared = 0;
-    let min = 0;
-    let max = 0;
-    const pixels = ellipseGeometry.getCoordinates()[0].flat();
-    for (let i = 0; i < pixels.length; i++) {
-      sum += pixels[i];
-      sumSquared += pixels[i] * pixels[i];
-      min = Math.min(min, pixels[i]);
-      max = Math.max(max, pixels[i]);
-      count++;
-    }
-
-    const mean = sum / count;
-    const variance = sumSquared / count - mean * mean;
-    const stdDev = Math.sqrt(variance);
-
-    const value1 = `Area: ${area.toFixed(2)} ${unitSuffix}`;
-    const value2 = `Mean: ${mean.toFixed(2)} Std Dev:${stdDev.toFixed(2)}`;
-    const value = `${value1}\n${value2}`;
-    markupManager.update({
-      feature: ellipseFeature,
-      value,
-      coordinate: ellipseGeometry.getLastCoordinate(),
-    });
-  }
-};
+import onSetFeatureStyle from "./onSetFeatureStyle";
+import getMeasurements from "./getMeasurements";
+import updateMarkup from "./updateMarkup";
+import getEllipsePolygonFromMovingLine from "./getEllipsePolygonFromMovingLine";
+import createAndAddEllipseFeature from "./createAndAddEllipseFeature";
 
 const isDrawingEllipse = (drawingOptions) =>
   drawingOptions.geometryType === Enums.GeometryType.Ellipse;
@@ -140,328 +18,273 @@ const drawEllipse = (
   currentPoint,
   viewerProperties
 ) => {
-  const { drawingSource, markupManager } = viewerProperties;
-
-  function midpoint(x1, x2, y1, y2) {
-    return [(x1 + x2) / 2, (y1 + y2) / 2];
-  }
-
-  const mid = midpoint(
-    fixedPoint[0],
-    movingPoint[0],
-    fixedPoint[1],
-    movingPoint[1]
-  );
+  const { drawingSource } = viewerProperties;
 
   const ellipseId = getEllipseId(feature);
-  const ellipse = drawingSource.getFeatureById(ellipseId);
+  const ellipseFeature = drawingSource.getFeatureById(ellipseId);
 
-  const center = mid;
-  const last = currentPoint;
-  const dx = center[0] - last[0];
-  const dy = center[1] - last[1];
-  let radius = Math.sqrt(dx * dx + dy * dy);
-  radius = radius > 0 ? radius : Number.MIN_SAFE_INTEGER;
-  const circle = new CircleGeometry(center, radius);
-  const polygon = fromCircle(circle);
-  polygon.scale(dx / radius, dy / radius);
+  const ellipsePolygon = getEllipsePolygonFromMovingLine(
+    fixedPoint,
+    movingPoint,
+    currentPoint
+  );
 
-  if (!ellipse) {
-    const geometry = polygon;
-    const ellipseFeature = new Feature({ geometry });
-    ellipseFeature.setId(ellipseId);
-    ellipseFeature.setProperties(
-      {
-        isEllipseShape: true,
-        [Enums.InternalProperties.ReadOnly]: true,
-        subFeatures: [feature],
-      },
-      true
-    );
-    feature.setProperties(
-      {
-        [Enums.InternalProperties.CantBeTranslated]: true,
-      },
-      true
-    );
-    ellipseFeature.setStyle(getEllipseStyle());
-
-    /** Remove markup from handles to add a new one to ellipse */
-    markupManager.remove(feature.getId());
-    markupManager.create({
-      feature: ellipseFeature,
-      style: feature.get(Enums.InternalProperties.StyleOptions),
-    });
-
-    drawingSource.addFeature(ellipseFeature);
+  if (!ellipseFeature) {
+    createAndAddEllipseFeature(feature, ellipsePolygon, viewerProperties);
   } else {
-    const geometry = ellipse.getGeometry();
-    geometry.setCoordinates(polygon.getCoordinates());
+    const geometry = ellipseFeature.getGeometry();
+    geometry.setCoordinates(ellipsePolygon.getCoordinates());
   }
 };
 
-const getChangeEvent = (viewerProperties) => {
-  const { map, drawingSource } = viewerProperties;
-  /**
-   * This is used to avoid changing features while
-   * dragging because of getClosestFeatureToCoordinate call
-   */
-  let draggedFeature,
-    draggedHandle = null;
-  map.on(Enums.MapEvents.POINTER_UP, () => {
-    draggedFeature = null;
-    draggedHandle = null;
-  });
+/**
+ * This is used to avoid changing features while
+ * dragging because of getClosestFeatureToCoordinate call
+ */
+let draggedFeature = null;
+let draggedHandle = null;
 
-  return (event) => {
-    const handleCoordinate = event.coordinate;
-    const handle = { x: handleCoordinate[0], y: handleCoordinate[1] };
+const resetEventState = () => {
+  draggedFeature = null;
+  draggedHandle = null;
+};
 
-    if (draggedFeature === null) {
-      draggedFeature =
-        drawingSource.getClosestFeatureToCoordinate(handleCoordinate);
-    }
+const global = {};
 
-    if (!draggedFeature) {
-      return;
-    }
+const onPointerDragHandler = (event) => {
+  const { viewerProperties } = global;
+  const { drawingSource } = viewerProperties;
 
-    const { isEllipseHandles } = draggedFeature.getProperties();
+  const handleCoordinate = event.coordinate;
+  const handle = { x: handleCoordinate[0], y: handleCoordinate[1] };
 
-    if (!isEllipseHandles) {
-      return;
-    }
+  if (draggedFeature === null) {
+    draggedFeature =
+      drawingSource.getClosestFeatureToCoordinate(handleCoordinate);
+  }
 
-    const draggedFeatureGeometry = draggedFeature.getGeometry();
-    const coords = draggedFeatureGeometry.getCoordinates();
+  if (!draggedFeature) {
+    return;
+  }
 
-    /** Disable dragging handle line */
-    if (
-      handleCoordinate[0] !== coords[0][0] &&
-      handleCoordinate[1] !== coords[0][1] &&
-      handleCoordinate[0] !== coords[1][0] &&
-      handleCoordinate[1] !== coords[1][1]
-    ) {
-      return;
-    }
+  const { isEllipseHandles } = draggedFeature.getProperties();
 
-    if (draggedHandle === null) {
-      const start = { x: coords[0][0], y: coords[0][1] };
-      const end = { x: coords[1][0], y: coords[1][1] };
-      const distanceStart = distance(handle, start);
-      const distanceEnd = distance(handle, end);
-      draggedHandle = distanceStart < distanceEnd ? "start" : "end";
-    }
+  if (!isEllipseHandles) {
+    return;
+  }
 
-    if (draggedHandle === "start") {
-      drawEllipse(
-        draggedFeature,
-        coords[0],
-        coords[1],
-        handleCoordinate,
-        viewerProperties
-      );
-      updateMarkup(draggedFeature, viewerProperties);
-      return;
-    }
+  const draggedFeatureGeometry = draggedFeature.getGeometry();
+  const draggedGeomCoords = draggedFeatureGeometry.getCoordinates();
 
-    if (draggedHandle === "end") {
-      drawEllipse(
-        draggedFeature,
-        coords[1],
-        coords[0],
-        handleCoordinate,
-        viewerProperties
-      );
-      updateMarkup(draggedFeature, viewerProperties);
-      return;
-    }
+  /** Disable dragging handle line */
+  if (
+    handleCoordinate[0] !== draggedGeomCoords[0][0] &&
+    handleCoordinate[1] !== draggedGeomCoords[0][1] &&
+    handleCoordinate[0] !== draggedGeomCoords[1][0] &&
+    handleCoordinate[1] !== draggedGeomCoords[1][1]
+  ) {
+    return;
+  }
+
+  if (draggedHandle === null) {
+    const start = { x: draggedGeomCoords[0][0], y: draggedGeomCoords[0][1] };
+    const end = { x: draggedGeomCoords[1][0], y: draggedGeomCoords[1][1] };
+    const distanceStart = distance(handle, start);
+    const distanceEnd = distance(handle, end);
+    draggedHandle = distanceStart < distanceEnd ? "start" : "end";
+  }
+
+  if (draggedHandle === "start") {
+    drawEllipse(
+      draggedFeature,
+      draggedGeomCoords[0],
+      draggedGeomCoords[1],
+      handleCoordinate,
+      viewerProperties
+    );
+    return;
+  }
+
+  if (draggedHandle === "end") {
+    drawEllipse(
+      draggedFeature,
+      draggedGeomCoords[1],
+      draggedGeomCoords[0],
+      handleCoordinate,
+      viewerProperties
+    );
+    return;
+  }
+};
+
+const attachChangeEvents = (ellipseHandlesFeature, viewerProperties) => {
+  const { map, markupManager } = viewerProperties;
+
+  ellipseHandlesFeature.setProperties(
+    {
+      isEllipseHandles: true,
+      [Enums.InternalProperties.IsSilentFeature]: true,
+    },
+    true
+  );
+
+  const ellipseHandlesGeometry = ellipseHandlesFeature.getGeometry();
+
+  const onEllipseHandlesGeometryChange = () => {
+    const [first, last] = ellipseHandlesGeometry.getCoordinates();
+    drawEllipse(ellipseHandlesFeature, first, last, last, viewerProperties);
+    /** We use a different format than markup manager builtin change event. */
+    updateMarkup(ellipseHandlesFeature, viewerProperties);
   };
+
+  ellipseHandlesGeometry.on(
+    Enums.FeatureGeometryEvents.CHANGE,
+    onEllipseHandlesGeometryChange
+  );
+
+  // const interactions = map.getInteractions();
+  // const draw = interactions.getArray().find((i) => i instanceof Draw);
+  // if (draw) {
+  //   const onDrawEndHandler = () => {
+  //     ellipseHandlesGeometry.un(
+  //       Enums.FeatureGeometryEvents.CHANGE,
+  //       onEllipseHandlesGeometryChange
+  //     );
+  //     draw.un(Enums.InteractionEvents.DRAW_END, onDrawEndHandler);
+  //   };
+  //   draw.on(Enums.InteractionEvents.DRAW_END, onDrawEndHandler);
+  // }
+};
+
+const attachPointerEvents = (viewerProperties) => {
+  const { map } = viewerProperties;
+  global.viewerProperties = viewerProperties;
+  map.on(Enums.MapEvents.POINTER_UP, resetEventState);
+  map.on(Enums.MapEvents.POINTER_DRAG, onPointerDragHandler);
 };
 
 const ellipse = Object.assign({}, annotationInterface, {
   onAdd: (feature, viewerProperties) => {
-    const { markupManager, drawingSource } = viewerProperties;
-    const { isEllipseHandles } = feature.getProperties();
-    if (isEllipseHandles) {
-      /** Remove markup from handles to add a new one to ellipse */
-      markupManager.remove(feature.getId());
-      const ellipse = drawingSource.getFeatureById(getEllipseId(feature));
-      if (ellipse) {
-        markupManager.create({
-          feature: ellipse,
-          style: feature.get(Enums.InternalProperties.StyleOptions),
-        });
-      }
-    }
+    // const { markupManager, drawingSource } = viewerProperties;
+    // const { measurements, isEllipseShape, isEllipseHandles } =
+    //   feature.getProperties();
+    // if (isEllipseHandles) {
+    //   markupManager.remove(feature.getId());
+    // }
+    // /** TODO: Remove area code check. */
+    // if (
+    //   measurements &&
+    //   JSON.stringify(measurements).includes("G-D7FE") &&
+    //   !isEllipseShape &&
+    //   !isEllipseHandles
+    // ) {
+    //   const ellipseFeature = feature;
+    //   const ellipseHandlesFeatureId = ellipseFeature.getId();
+    //   const ellipseId = getEllipseId(ellipseFeature);
+    //   ellipseFeature.setId(ellipseId);
+    //   ellipseFeature.setProperties(
+    //     {
+    //       isEllipseShape: true,
+    //       [Enums.InternalProperties.ReadOnly]: true,
+    //     },
+    //     true
+    //   );
+    //   const ellipseHandlesFeature = drawingSource.getFeatureById(
+    //     ellipseHandlesFeatureId
+    //   );
+    //   if (!ellipseHandlesFeature) {
+    //     const [coords] = ellipseFeature.getGeometry().getCoordinates();
+    //     const middle = Math.round(coords.length / 2);
+    //     const padding = 700; /** TODO: Calculate based on radius */
+    //     const paddedStart = [coords[0][0] + padding, coords[0][1] + padding];
+    //     const paddedEnd = [
+    //       coords[middle][0] - padding,
+    //       coords[middle][1] - padding,
+    //     ];
+    //     const handles = new LineString([paddedStart, paddedEnd]);
+    //     const ellipseHandlesFeature = new Feature({ geometry: handles });
+    //     ellipseHandlesFeature.setId(ellipseHandlesFeatureId);
+    //     ellipseHandlesFeature.setProperties(
+    //       {
+    //         isEllipseHandles: true,
+    //         [Enums.InternalProperties.IsSilentFeature]: true,
+    //       },
+    //       true
+    //     );
+    //     ellipseHandlesFeature.setStyle(ellipseHandlesStyleFunction);
+    //     ellipseHandlesFeature.set(
+    //       Enums.InternalProperties.StyleOptions,
+    //       styles
+    //     );
+    //     ellipseFeature.setProperties(
+    //       {
+    //         subFeatures: [ellipseHandlesFeature],
+    //       },
+    //       true
+    //     );
+    //     drawingSource.addFeature(ellipseHandlesFeature);
+    //     attachChangeEvents(ellipseHandlesFeature, viewerProperties);
+    //     attachPointerEvents(viewerProperties);
+    //   }
+    // }
+    // return true;
   },
-  onInit: (viewerProperties) => {},
   onDrawEnd: (event, viewerProperties) => {
-    const { map } = viewerProperties;
-    map.on(Enums.MapEvents.POINTER_DRAG, getChangeEvent(viewerProperties));
+    global.viewerProperties = viewerProperties;
+    attachPointerEvents(viewerProperties);
   },
   onDrawStart: (event, viewerProperties) => {
-    const { drawingOptions, map } = viewerProperties;
+    const { drawingOptions } = viewerProperties;
     if (isDrawingEllipse(drawingOptions)) {
       const ellipseHandlesFeature = event.feature;
-
-      ellipseHandlesFeature.setProperties(
-        {
-          isEllipseHandles: true,
-          [Enums.InternalProperties.IsSilentFeature]: true,
-        },
-        true
-      );
-
-      // TODO: Add built in highlighting of features
-      // let selected = null;
-      // map.on('pointermove', function (e) {
-      //   if (selected !== null) {
-      //     selected.setStyle(undefined);
-      //     selected = null;
-      //   }
-
-      //   map.forEachFeatureAtPixel(e.pixel, function (f) {
-      //     selected = f;
-      //     // add highlight
-      //     // f.setStyle(highlightStyle);
-      //     return true;
-      //   });
-
-      //   if (selected) {
-      //     // add highlight
-      //   } else {
-      //     // remove highlight
-      //   }
-      // });
-
-      ellipseHandlesFeature.getGeometry().on("change", (event) => {
-        const geometry = event.target;
-        const [first, last] = geometry.getCoordinates();
-        drawEllipse(ellipseHandlesFeature, first, last, last, viewerProperties);
-        updateMarkup(ellipseHandlesFeature, viewerProperties);
-      });
+      attachChangeEvents(ellipseHandlesFeature, viewerProperties);
     }
   },
   onRemove: (feature, { drawingSource, markupManager }) => {
-    const { isEllipse } = feature.getProperties();
-    if (isEllipse) {
-      const ellipseHandlesFeatureId = getEllipseHandlesId(feature);
-      const ellipseHandlesFeature = drawingSource.getFeatureById(
-        ellipseHandlesFeatureId
-      );
-      drawingSource.removeFeature(ellipseHandlesFeature);
-      markupManager.remove(ellipseHandlesFeature.getId());
-    }
-  },
-  getMeasurements: (feature, viewerProperties) => {
-    const { drawingSource, pyramid } = viewerProperties;
-
-    const { isEllipseHandles, isEllipseShape } =
-      feature.getProperties();
-    const isEllipse = isEllipseHandles || isEllipseShape;
-
-    if (isEllipse) {
-      let ellipseHandlesFeature;
-      let ellipseFeature;
-      let points = [];
-      let area = 0;
-
-      if (ellipseHandlesFeature) {
-        ellipseHandlesFeature = feature;
-
-        const ellipseFeatureId = getEllipseId(ellipseHandlesFeature);
-        ellipseFeature = drawingSource.getFeatureById(ellipseFeatureId);
-
-        if (ellipseFeature) {
-          area = getFeatureScoord3dArea(ellipseFeature, pyramid);
-          points = ellipseFeature.getGeometry().getCoordinates().map(coordinate => ({
-            x: coordinate[0],
-            y: coordinate[1] 
-          }));
-        }
-      }
-
-      if (isEllipseShape) {
-        ellipseFeature = feature;
-        area = getFeatureScoord3dArea(ellipseFeature, pyramid);
-        points = ellipseFeature.getGeometry().getCoordinates().map(coordinate => ({
-          x: coordinate[0],
-          y: coordinate[1] 
-        }));
-      }
-
-      const ellipse = new dcmjs.utilities.TID300.Ellipse({
-        area,
-        points
-      });
-
-      const contentItem = ellipse.contentItem();
-      const numContentItems = contentItem.filter(
-        (ci) => ci.ValueType === "NUM"
-      );
-      return numContentItems.map((measurement) => {
-        /** Update format wrong in dcmjs */
-        measurement.ConceptNameCodeSequence = [
-          measurement.ConceptNameCodeSequence,
-        ];
-        return measurement;
-      });
-    }
-    return [];
-  },
-  onSetFeatureStyle: (feature, styleOptions, viewerProperties) => {
-    const { drawingSource } = viewerProperties;
-    const { isEllipseHandles, isEllipse, isEllipseShape } =
-      feature.getProperties();
-
-    const isEllipseAnnotation = isEllipseHandles || isEllipse || isEllipseShape;
-
-    if (!isEllipseAnnotation) {
-      return;
-    }
-
-    if (isEllipseAnnotation) {
-      if (styleOptions.stroke) {
-        styles.stroke = Object.assign(styles.stroke, styleOptions.stroke);
-      }
-
-      if (styleOptions.image) {
-        styles.image = Object.assign(styles.image, styleOptions.image);
-      }
-    }
-
-    const ellipseStyle = getEllipseStyle();
-
-    if (isEllipseHandles || isEllipse) {
-      feature.setStyle(ellipseHandlesStyleFunction);
-      feature.set(Enums.InternalProperties.StyleOptions, styles);
-
+    const { isEllipseHandles, isEllipseShape } = feature.getProperties();
+    if (isEllipseHandles) {
       const ellipseFeatureId = getEllipseId(feature);
       const ellipseFeature = drawingSource.getFeatureById(ellipseFeatureId);
       if (ellipseFeature) {
-        ellipseFeature.setStyle(ellipseStyle);
-        ellipseFeature.set(Enums.InternalProperties.StyleOptions, styles);
+        drawingSource.removeFeature(ellipseFeature);
       }
     }
 
     if (isEllipseShape) {
-      feature.setStyle(ellipseStyle);
-      feature.set(Enums.InternalProperties.StyleOptions, styles);
-
       const ellipseHandlesFeatureId = getEllipseHandlesId(feature);
       const ellipseHandlesFeature = drawingSource.getFeatureById(
         ellipseHandlesFeatureId
       );
       if (ellipseHandlesFeature) {
-        ellipseHandlesFeature.setStyle(ellipseHandlesStyleFunction);
-        ellipseHandlesFeature.set(
-          Enums.InternalProperties.StyleOptions,
-          styles
-        );
+        drawingSource.removeFeature(ellipseHandlesFeature);
+        markupManager.remove(ellipseHandlesFeature.getId());
       }
     }
   },
-})
+  getMeasurements,
+  onSetFeatureStyle,
+});
 
 export default ellipse;
+
+// TODO: Add built in highlighting of features
+// let selected = null;
+// map.on('pointermove', function (e) {
+//   if (selected !== null) {
+//     selected.setStyle(undefined);
+//     selected = null;
+//   }
+
+//   map.forEachFeatureAtPixel(e.pixel, function (f) {
+//     selected = f;
+//     // add highlight
+//     // f.setStyle(highlightStyle);
+//     return true;
+//   });
+
+//   if (selected) {
+//     // add highlight
+//   } else {
+//     // remove highlight
+//   }
+// });
